backend/requirements.txt
bson~=0.5.10
pymongo~=4.9.2
python-dotenv~=1.0.1
Flask~=3.1.0
Flask-Cors~=5.0.0